fix: preserve fractional executedQty as Decimal instead of int
order.py: int(float("0.001")) == 0 for BTC/ETH step sizes — replaced
with Decimal(str(...)) in both place_order_via (line 167) and
list_trades_via (line 236). OrderResponse.filled_quantity type changed
from int to Decimal(default "0"); BinanceTradeStatus.deal_quantity and
cancel_quantity changed from Optional[int] to Optional[Decimal].

Downstream consumers (lcz-quant etc.) must update any int assumptions
on deal_quantity.

+++ binance_shioaji_sdk/trade.py
@@ -37,9 +37,9 @@ class BinanceTradeStatus:
     status: BinanceOrderStatusEnum
     status_code: str
     order_datetime: str                       # ISO 8601 UTC
-    deal_quantity: Optional[int] = None       # cumulative filled qty (Binance: executedQty)
+    deal_quantity: Optional[Decimal] = None    # cumulative filled qty (Binance: executedQty)
     order_quantity: Optional[Decimal] = None  # original submitted qty (Binance: origQty)
-    cancel_quantity: Optional[int] = None     # qty remaining after cancellation
+    cancel_quantity: Optional[Decimal] = None  # qty remaining after cancellation
     modified_price: float = 0.0               # weighted avg fill price (0.0 if no fill)
     msg: str = ""                             # broker message (error reason on failure)
